test: move PresentedSurfaceFixture into its own file
